Fix example notebooks, black, increment version.
