+++ src/xradio/vis/read_processing_set.py
@@ -8,7 +8,7 @@ import s3fs
 
 def read_processing_set(
     ps_store: str,
-    intents: list = None,
+    obs_modes: list = None,
 ) -> processing_set:
     """Creates a lazy representation of a Processing Set (only meta-data is loaded into memory).
 
@@ -16,9 +16,9 @@ def read_processing_set(
     ----------
     ps_store : str
         String of the path and name of the processing set. For example '/users/user_1/uid___A002_Xf07bba_Xbe5c_target.lsrk.vis.zarr'.
-    intents : list, optional
-        A list of the intents to be read for example ['OBSERVE_TARGET#ON_SOURCE']. The intents in a processing set can be seem by calling processing_set.summary().
-        By default None, which will read all intents.
+    obs_modes : list, optional
+        A list of obs_mode to be read for example ['OBSERVE_TARGET#ON_SOURCE']. The obs_mode in a processing set can be seem by calling processing_set.summary().
+        By default None, which will read all obs_mode.
 
     Returns
     -------
@@ -37,7 +37,9 @@ def read_processing_set(
         xds = _open_dataset(ms_main_store, file_system)
         data_groups = xds.attrs["data_groups"]
 
-        if (intents is None) or (xds.attrs["partition_info"]["intent"] in intents):
+        if (obs_modes is None) or (
+            xds.attrs["partition_info"]["obs_mode"] in obs_modes
+        ):
             sub_xds_dict, field_and_source_xds_dict = _read_sub_xds(
                 ms_store, file_system=file_system, data_groups=data_groups
             )
